refactor : 로또 번호 생성 로직을 DataSource와 Repository로 분리
