+++ src/main/kotlin/lotto/data/datasource/LottoDataSource.kt
@@ -1,4 +1,4 @@
-package lotto.data
+package lotto.data.datasource
 
 import java.util.TreeSet
