+++ src/main/kotlin/lotto/domain/repository/LottoRepository.kt
@@ -1,0 +1,7 @@
+package lotto.domain.repository
+
+import java.util.TreeSet
+
+interface LottoRepository {
+    fun generateLottoNumbers(): TreeSet<Int>
+}
