+++ src/test/kotlin/lotto/LottoTest.kt
@@ -1,6 +1,7 @@
 package lotto
 
 import lotto.app.DependencyInjector
+import lotto.data.datasource.generateLotto
 import lotto.domain.enums.Rank
 import lotto.domain.lotto.Lotto
 import org.junit.jupiter.api.Assertions.assertEquals
@@ -39,7 +40,7 @@ class LottoTest {
 
     @Test
     fun `로또 번호 반환 테스트`() {
-        val lottoFactory = dependencyInjector.injectLottoDataSource()
+        val lottoFactory = generateLotto()
         val lotto = lottoFactory()
         repeat(100000) {
             assertEquals(lotto.size, 6)
